04-08_파이썬과제/sss일잘러/requirements.txt
colorama==0.4.6
docx2pdf==0.1.8
et-xmlfile==1.1.0
Faker==24.8.0
lxml==5.2.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-pptx==0.6.23
pytz==2024.1
pywin32==306
six==1.16.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
XlsxWriter==3.2.0
